Cleanup (java decompiler test optimization)
