+++ test/org/jetbrains/java/decompiler/BulkDecompilationTest.java
